The refusal constructors move out of the write planner (refusal-constructor-home)
A pure move, and nothing else. `plan.ts` — the write planner — owned the
refusal sentences this layer says NO with, because the planner is where a
refusal is usually raised and so is where each of them was born. But three of
them are said by `query.ts` too, which is pure reads and plans nothing, and it
was importing them from the writer to say what a write says.

They live in `refusals.ts` now, and both files import from it. Not one of them
takes a planning scope; not one of them knows a request exists.

Named by #320's altitude review and deferred there, since #320 was about
batching MCP reads and a ~220-line relocation of code it did not author would
have made that diff harder to read for what it was about.
